// src/global/candidat.d.ts
export interface candidat{
  "noCandidat": string,
  "adresse": string,
  "codePostal": string,
  "confirmationCandidat": string,
  "dateNaissance": string,
  "dateReponseCandidat": string,
  "email": string,
  "lieuNaissance": string,
  "listeSelection": string,
  "mobile":string,
  "nationalite": string,
  "nom": string,
  "paysOrigine": string,
  "prenom": string,
  "selectionNoOrdre":string,
  "sexe": string,
  "telephone":string,
  "universiteOrigine": string,
  "ville":string,
}